Untangle frontend module dependencies
