--- client/src/canvas/types.ts
@@ -1,1 +1,0 @@
-export type Context2D = CanvasRenderingContext2D;
