--- client/src/game/elements.ts
@@ -1,5 +1,0 @@
-import $ from "jquery";
-
-export const gameElements = {
-    canvas: $<HTMLCanvasElement>("#game-box"),
-};
