--- client/src/messaging/elements.ts
@@ -1,7 +1,0 @@
-import $ from "jquery";
-
-export const messagingElements = {
-    devTools: $("#dev-tools"),
-    openCloseDevTools: $("#open-close-dev-tools"),
-    messageForm: $<HTMLFormElement>("#message-form"),
-};
