--- client/src/networking/elements.ts
@@ -1,11 +1,0 @@
-import $ from "jquery";
-
-export const networkElements = {
-    connectedBox: $("#connected-box"),
-    errorBox: $("#error-box"),
-    messagesStats: $("#messages-stats"),
-    messagesBox: $("#messages-box"),
-    viteMode: $("#vite-mode"),
-    joinForm: $<HTMLFormElement>("#join-form"),
-    inactiveOverlay: $(".inactive-overlay"),
-};
