+++ client/src/ui/gameConstants.ts
