+++ server/src/main/java/io/github/aggarcia/clients/updates/CreateFirstPlayer.java
@@ -27,7 +27,8 @@ public record CreateFirstPlayer(
     @Override
     public Optional<byte[]> reply() {
         var reply = JoinReply.newBuilder().setServerId(serverId);
-        var wrappedReply = SocketMessage.newBuilder().setJoinReply(reply).build();
+        var wrappedReply =
+            SocketMessage.newBuilder().setJoinReply(reply).build();
         return Optional.of(serialize(wrappedReply));
     }
 }
